examples/vad_implementation/requirements.txt
deepgram_sdk==3.4.0
librosa==0.10.2.post1
numpy==2.1.1
torch==2.4.0
